Refactor JavaRuntimeFinder into subclasses & clean up
The getAvailableRuntimes function is now parseable, & the new
PlatformRuntimeFinder interface makes it clear what each stage does.
