+++ launcher/src/main/java/com/skcraft/launcher/dialog/ConfigurationDialog.java
@@ -9,9 +9,9 @@ package com.skcraft.launcher.dialog;
 import com.skcraft.launcher.Configuration;
 import com.skcraft.launcher.Launcher;
 import com.skcraft.launcher.dialog.component.BetterComboBox;
-import com.skcraft.launcher.launch.AddJavaRuntime;
-import com.skcraft.launcher.launch.JavaRuntime;
-import com.skcraft.launcher.launch.JavaRuntimeFinder;
+import com.skcraft.launcher.launch.runtime.AddJavaRuntime;
+import com.skcraft.launcher.launch.runtime.JavaRuntime;
+import com.skcraft.launcher.launch.runtime.JavaRuntimeFinder;
 import com.skcraft.launcher.persistence.Persistence;
 import com.skcraft.launcher.swing.*;
 import com.skcraft.launcher.util.SharedLocale;
