+++ launcher/src/main/java/com/skcraft/launcher/dialog/InstanceSettingsDialog.java
@@ -3,9 +3,9 @@ package com.skcraft.launcher.dialog;
 import com.skcraft.launcher.Instance;
 import com.skcraft.launcher.InstanceSettings;
 import com.skcraft.launcher.dialog.component.BetterComboBox;
-import com.skcraft.launcher.launch.JavaRuntime;
-import com.skcraft.launcher.launch.JavaRuntimeFinder;
 import com.skcraft.launcher.launch.MemorySettings;
+import com.skcraft.launcher.launch.runtime.JavaRuntime;
+import com.skcraft.launcher.launch.runtime.JavaRuntimeFinder;
 import com.skcraft.launcher.persistence.Persistence;
 import com.skcraft.launcher.swing.FormPanel;
 import com.skcraft.launcher.swing.LinedBoxPanel;
