+++ launcher/src/main/java/com/skcraft/launcher/launch/runtime/AddJavaRuntime.java
@@ -1,4 +1,4 @@
-package com.skcraft.launcher.launch;
+package com.skcraft.launcher.launch.runtime;
 
 import java.io.File;
